pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["kaori_agent*"]

[project]
name = "kaori-agent"
version = "0.1.0"
description = "Personal assistant agent with tool_use support"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.43",
    "openai>=1.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "mcp>=1.0",
]

[project.optional-dependencies]
cli = [
    "rich>=13.0",
    "prompt_toolkit>=3.0",
]
server = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "websockets>=13.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
]
